Redirects users to auth automatically; we can therefore assume that users are authed if they're viewing a route on the site
